[fix]rails_helperに記述していたものを分割、specのログイン修正
