+++ spec/rails_helper.rb
@@ -8,6 +8,7 @@ require File.expand_path('../config/environment', __dir__)
 abort('The Rails environment is running in production mode!') if Rails.env.production?
 require 'rspec/rails'
 require 'capybara/rspec'
+
 # Add additional requires below this line. Rails is not loaded until this point!
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
@@ -23,8 +24,7 @@ require 'capybara/rspec'
 # directory. Alternatively, in the individual `*_spec.rb` files, manually
 # require only the support files necessary.
 #
-# Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
-Dir[Rails.root.join("spec/support/*.rb")].each { |f| require f }
+Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
 # Checks for pending migrations and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove these lines.
 begin
@@ -32,25 +32,6 @@ begin
 rescue ActiveRecord::PendingMigrationError => e
   puts e.to_s.strip
   exit 1
-end
-Capybara.configure do |config|
-  config.server_host = Socket.ip_address_list.detect(&:ipv4_private?).ip_address
-  config.server_port = 3001
-  config.default_driver    = :selenium_chrome
-  config.javascript_driver = :selenium_chrome
-  config.default_max_wait_time = 5
-  config.ignore_hidden_elements = true
-end
-Capybara.register_driver :selenium_chrome do |app|
-  url = "http://chrome:4444/wd/hub"
-  opts = { desired_capabilities: :chrome, browser: :remote, url: url }
-  # driver = Capybara::Selenium::Driver.new(app, opts)
-  Capybara::Selenium::Driver.new(app, opts)
-end
-def wait_until(wait_time = Capybara.default_max_wait_time)
-  Timeout.timeout(wait_time) do
-    loop until yield
-  end
 end
 
 include Warden::Test::Helpers
@@ -62,6 +43,7 @@ RSpec.configure do |config|
   # If you're not using ActiveRecord, or you'd prefer not to run each of your
   # examples within a transaction, remove the following line or assign false
   # instead of true.
+  config.include SystemSupport, type: :system
   config.use_transactional_fixtures = true
   config.before(:each, type: :system) do
     driven_by Capybara.default_driver
