+++ spec/system/tweets_spec.rb
@@ -1,22 +1,15 @@
 require 'rails_helper'
 
 RSpec.feature 'Tweets', type: :system, js: true do
-  #   find(".fa-send-o").click
-  #   expect{
-  #     fill_in "tweet_body", with: "本文テスト"
-  #     click_button "投稿する"
+  let!(:member) { FactoryBot.create(:member) }
 
-  #     expect(page).to have_content '投稿しました。'
-  #     expect(page).to have_content '本文テスト'
-  #     expect(page).to have_content "#{member.name}"
-
-  #   }.to change(member.tweets, :count).by(1)
-  # end
   before {
-    member = FactoryBot.create(:member)
-    login_as(member, scope: :member, :run_callbacks => false)
-    visit root_path
-    wait_until { page.has_css?("h1", text: "Infomus") }
+    token = member.confirmation_token
+    visit member_confirmation_path(confirmation_token: token)
+    fill_in 'member_email', with: member.email
+    fill_in 'member_password', with: member.password
+    page.save_screenshot "ss.png"
+    click_button "ログイン"
   }
   scenario ' ユーザーが新しいツイートを作成する ' do
     click_link(class: 'post_tweet')
@@ -30,13 +23,4 @@ RSpec.feature 'Tweets', type: :system, js: true do
       expect(page).to have_content "#{member.name}"
     }.to change(member.tweets, :count).by(1)
   end
-  # context ' ユーザーが新しいツイートを作成する ' do
-  #   it {
-  #     fill_in "tweet_body", with: "本文テスト"
-  #     click_button "投稿する"
-  #     expect(page).to have_content '投稿しました。'
-  #     expect(page).to have_content '本文テスト'
-  #     expect(page).to have_content "#{member.name}"
-  #   }
-  #   end
 end
